Refactored gnirsSort.py
