+++ gnirs.cfg
@@ -6,9 +6,16 @@ niftyVersion = "1.0.1"
 manualMode = False
 overwrite = False
 scienceOneDExtraction = True
-scienceDirectoryList = ['/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/ShortBlue_SXD_32_0.68arcsec_1.6499⁩/obsGN_2011A_Q_126_6⁩']
-telluricDirectoryList = ['/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/ShortBlue_SXD_32_0.68arcsec_1.6499⁩/Tellurics/obsGN_2011A_Q_126_13']
-calibrationDirectoryList = ['/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/Calibrations_ShortBlue_SXD_32_0.68arcsec_1.6499⁩']
+
+#scienceDirectoryList = ['/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/ShortBlue_SXD_32_0.68arcsec_1.6499⁩/obsGN_2011A_Q_126_6⁩']
+#telluricDirectoryList = ['/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/ShortBlue_SXD_32_0.68arcsec_1.6499⁩/Tellurics/obsGN_2011A_Q_126_13']
+#calibrationDirectoryList = ['/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/Calibrations_ShortBlue_SXD_32_0.68arcsec_1.6499⁩']
+
+[ScienceDirectories]
+/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/ShortBlue_SXD_32_0.68arcsec_1.6499um/Science/⁩GN_2011A_Q_126_6⁩ = True
+
+[TelluricDirectories]
+/Users/vkhatu/GitHub/GNIRS_PIPELINE/NGC4736/20110516/ShortBlue_SXD_32_0.68arcsec_1.6499⁩/Tellurics/obsGN_2011A_Q_126_13 = True
 
 [gnirsPipelineConfig]
 getdata = False              ; Download the required GNIRS data
@@ -26,8 +33,6 @@ fluxCalibrationMethod = "gnirs"
 [dataConfig]
 rawPath = rawData/
 program = ""
-
-[headersConfig]
 
 [sortConfig]
 rawPath = rawData/
